# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "Sandbox Connect API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/v1/admin/outputs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "output-admin"
                ],
                "summary": "List outputs pending NHA approval",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum outputs (1-200)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.PendingOutputsResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error403"
                        }
                    }
                }
            }
        },
        "/v1/admin/outputs/{output_id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Queues idempotent Files Connect publication into the owner's object-backed workspace.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "output-admin"
                ],
                "summary": "Approve and publish output files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Output ID",
                        "name": "output_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique request key",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ApprovalResponse"
                        }
                    }
                }
            }
        },
        "/v1/admin/outputs/{output_id}/files": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "output-admin"
                ],
                "summary": "List CSV files for an output awaiting approval",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Output ID",
                        "name": "output_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.OutputFilesResponse"
                        }
                    }
                }
            }
        },
        "/v1/admin/outputs/{output_id}/files/{file_id}/preview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "output-admin"
                ],
                "summary": "Preview a pending CSV output file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Output ID",
                        "name": "output_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stable file ID",
                        "name": "file_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sandbox-backend-service_internal_filesconnect.Preview"
                        }
                    }
                }
            }
        },
        "/v1/bookings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists CPU and GPU slot bookings for the current user with optional status filter and pagination.\nThe `status` filter accepts comma-separated lifecycle states: `scheduled`, `ready`, `active`, `shutting_down`, `completed`, `cancelled`, `expired`, or `all` to disable filtering.\n`notebookUrl` is returned only when the booking is `active` and the linked notebook resource has been applied.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "List user bookings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated status filter, or 'all' to omit filtering",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max rows (default 10, max 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.BookingsListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error401"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error429"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error500"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a CPU or GPU booking from code-defined categories and slot templates selected by SLOT_CONFIG_PROFILE.\n\nNew bookings are inserted as `scheduled`. Lifecycle automation moves `scheduled` bookings to `ready` at slot start, then to `active` once the Kubeflow Notebook reports ready replicas, then to `shutting_down` near slot end, and finally to `completed` at slot end.\n\nBooking limits are enforced as two separate rules.\nMaxActiveBookings counts scheduled, ready, active, and shutting_down bookings for the same user and category. When reached, the API returns 400 with \"Active booking limit exceeded\".\nMaxBookingsPerWeek counts scheduled, ready, active, shutting_down, and completed bookings for the same user, category, and selected week. When reached, the API returns 400 with \"Weekly booking limit exceeded\".\nCancelled bookings do not count toward either limit. Expired bookings do not count toward weekly quota because expired can mean either no-show expiry after the booking became ready, or reset/cleanup of a stuck ready booking. Scheduled bookings reset before resources are ready become cancelled, not expired.\nThe previous \"You already have an upcoming booking\" restriction has been removed. Users may create multiple future bookings within MaxActiveBookings, MaxBookingsPerWeek, slot availability, and duplicate booking rules.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Create slot booking",
                "parameters": [
                    {
                        "description": "Booking request",
                        "name": "booking",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api.CreateBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.CreateBookingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error400"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error403"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error404"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error409"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error422"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error429"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error500"
                        }
                    }
                }
            }
        },
        "/v1/bookings/{id}/cancel": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancels a `scheduled` booking for the current user before resources are ready.\nCancel changes `scheduled` to `cancelled`. It does not operate on `ready`, `active`, or `shutting_down`; use terminate for those states.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Cancel scheduled booking",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.SwaggerMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error400"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error401"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error404"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error429"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error500"
                        }
                    }
                }
            }
        },
        "/v1/bookings/{id}/extend": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Extends an `active` booking to the next contiguous slot if that slot has capacity, the category has not reached its contiguous-slot limit, and the booking has not already been extended.\nThe booking remains `active`; `slot_keys`, `slot_end`, `shutdown_warning_sent_at`, and `extension_used` are updated so lifecycle timing follows the new end time.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Extend active booking by one slot",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ExtendBookingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error400"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error401"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error404"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error409"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error500"
                        }
                    }
                }
            }
        },
...
        "/v1/bookings/{id}/reset": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Resets only `scheduled` or `ready` bookings that are stuck or need cleanup.\n`scheduled` resets become `cancelled` and unlink any notebook id. `ready` resets become `expired`, set session and cleanup timestamps, unlink the notebook, and best-effort delete the linked Notebook/PVC.\nReset does not apply to `active`, `shutting_down`, `completed`, `cancelled`, or already `expired` bookings.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Reset stuck booking",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.SwaggerMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error400"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error401"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error500"
                        }
                    }
                }
            }
        },
        "/v1/bookings/{id}/terminate": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ends a `ready`, `active`, or `shutting_down` booking early and marks it `completed`.\nTerminate sets session and cleanup timestamps, unlinks the notebook, and best-effort deletes the linked Notebook/PVC. If the booking is already `completed`, the endpoint returns success without changing it.\nUse cancel or reset for `scheduled` bookings; terminate does not apply to `cancelled` or `expired` bookings.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "End booking session early",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.SwaggerMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error400"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error401"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error500"
                        }
                    }
                }
            }
        },
...
        "/v1/workspace/files": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "List approved output files in the user's workspace",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.WorkspaceFilesResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "cmd_api.ApprovalResponse": {
            "type": "object",
            "properties": {
                "outputId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "cmd_api.Error403": {
            "description": "Response for permission errors",
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Error message",
                    "type": "string",
                    "example": "string"
                },
                "title": {
                    "description": "Error title",
                    "type": "string",
                    "example": "Forbidden"
                },
                "type": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "cmd_api.PendingOutputsResponse": {
            "type": "object",
            "properties": {
                "outputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cmd_api.OutputResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}